requirements.txt
   pandas
   faker
   scikit-learn
   gradio
   groq
   python-dotenv
